refactor zdx docs layout
