+++ pkgs/community/zdx/docs/index.md
@@ -1,0 +1,4 @@
+# ZDX Documentation
+
+Welcome to the ZDX documentation.
+
