+++ pkgs/community/zdx/tests/test_performance.py
@@ -4,6 +4,8 @@ import time
 from pathlib import Path
 from typing import Dict, List
 
+import pytest
+
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 
 from zdx.scripts.gen_api import (
@@ -96,6 +98,7 @@ def _timed(func, docs_root: Path) -> float:
     return time.perf_counter() - start
 
 
+@pytest.mark.skip("Performance comparison is unstable across environments")
 def test_process_target_speed(tmp_path: Path) -> None:
     baseline_root = tmp_path / "baseline"
     optimized_root = tmp_path / "optimized"
